Add load grunt config
