+++ yeoman/grunt/bower.js
@@ -1,0 +1,5 @@
+module.exports = {
+  all: {
+    rjsConfig: '<%= yeoman.app %>/scripts/main.js'
+  }
+};
