+++ yeoman/grunt/clean.js
@@ -1,0 +1,10 @@
+module.exports = {
+  options: {
+    force: true
+  },
+  dist: ['.tmp', '<%= yeoman.dist %>/*'],
+  server: '.tmp',
+  aftercopy: [
+    '<%= yeoman.dist %>/*.html'
+  ]
+};
