+++ yeoman/grunt/csscomb.js
@@ -1,0 +1,8 @@
+module.exports = {
+  dist: {
+    expand: true,
+    cwd: 'app/styles/',
+    src: '*.scss',
+    dest: 'app/styles/'
+  }
+};
