+++ yeoman/grunt/cssmin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: {
+      '<%= yeoman.dist %>/main.css': [
+        '.tmp/styles/{,*/}*.css',
+        '<%= yeoman.app %>/styles/{,*/}*.css'
+      ]
+    }
+  }
+};
