+++ yeoman/grunt/imagemin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '<%= yeoman.app %>/images',
+      src: '{,*/}*.{png,jpg,jpeg}',
+      dest: '<%= yeoman.dist %>'
+    }]
+  }
+};
