+++ yeoman/grunt/jshint.js
@@ -1,0 +1,12 @@
+module.exports = {
+  options: {
+    jshintrc: '.jshintrc',
+    reporter: require('jshint-stylish')
+  },
+  all: [
+    'Gruntfile.js',
+    '<%= yeoman.app %>/scripts/{,*/}*.js',
+    '!<%= yeoman.app %>/scripts/vendor/*',
+    'test/spec/{,*/}*.js'
+  ]
+};
