+++ yeoman/grunt/open.js
@@ -1,0 +1,8 @@
+module.exports = {
+  server: {
+    path: 'http://localhost:<%= connect.options.port %>'
+  },
+  test: {
+    path: 'http://localhost:<%= connect.test.options.port %>'
+  }
+};
