+++ yeoman/grunt/uglify.js
@@ -1,0 +1,9 @@
+module.exports = {
+  dist: {
+    files: {
+      '<%= yeoman.dist %>/main.js': [
+        '.tmp/scripts/main.js'
+      ]
+    }
+  }
+};
