+++ yeoman/grunt/usemin.js
@@ -1,0 +1,7 @@
+module.exports = {
+  html: ['<%= yeoman.dist %>/{,*/}*.html'],
+  css: ['<%= yeoman.dist %>/{,*/}*.css'],
+  options: {
+    dirs: ['<%= yeoman.dist %>']
+  }
+};
