+++ yeoman/grunt/useminPrepare.js
@@ -1,0 +1,6 @@
+module.exports = {
+  html: '<%= yeoman.app %>/index.html',
+  options: {
+    dest: '<%= yeoman.dist %>'
+  }
+};
