+++ yeoman/grunt/yeoman.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt, options) {
+  return grunt.yeomanConfig;
+};
