+++ yeoman/package.json
@@ -28,6 +28,7 @@
     "grunt-usemin": "^2.4.0",
     "jit-grunt": "^0.9.1",
     "jshint-stylish": "^1.0.0",
+    "load-grunt-config": "^0.19.1",
     "postcss": "^5.0.14",
     "postcss-cssnext": "^2.4.0",
     "postcss-nested": "^1.0.0",
